LLM/requirements.txt
torch
transformers
sentence-transformers
torchvision
torchtext
torchsummary
torchmetrics
torchio
torchdiffeq
torch-geometric
fastapi
uvicorn
python-multipart
pillow
opencv-python
pyzbar
easyocr
numpy